Move `ex` files to directories correspoding to their modules names
