+++ benchmark/lib/benchee/merge_frames.exs
@@ -59,17 +59,17 @@ defmodule Membrane.VideoCompositor.Benchmark.MergeFrames do
       second: frame_4k
     }
 
-    {:ok, ffmpeg_internal_state_720p} = Membrane.VideoCompositor.FFmpeg.init(caps_720p)
-    {:ok, ffmpeg_internal_state_1080p} = Membrane.VideoCompositor.FFmpeg.init(caps_1080p)
-    {:ok, ffmpeg_internal_state_4k} = Membrane.VideoCompositor.FFmpeg.init(caps_4k)
+    {:ok, ffmpeg_internal_state_720p} = Membrane.VideoCompositor.Implementation.FFmpeg.init(caps_720p)
+    {:ok, ffmpeg_internal_state_1080p} = Membrane.VideoCompositor.Implementation.FFmpeg.init(caps_1080p)
+    {:ok, ffmpeg_internal_state_4k} = Membrane.VideoCompositor.Implementation.FFmpeg.init(caps_4k)
 
-    {:ok, opengl_cpp_internal_state_720p} = Membrane.VideoCompositor.OpenGL.Cpp.init(caps_720p)
-    {:ok, opengl_cpp_internal_state_1080p} = Membrane.VideoCompositor.OpenGL.Cpp.init(caps_1080p)
-    {:ok, opengl_cpp_internal_state_4k} = Membrane.VideoCompositor.OpenGL.Cpp.init(caps_4k)
+    {:ok, opengl_cpp_internal_state_720p} = Membrane.VideoCompositor.Implementation.OpenGL.Cpp.init(caps_720p)
+    {:ok, opengl_cpp_internal_state_1080p} = Membrane.VideoCompositor.Implementation.OpenGL.Cpp.init(caps_1080p)
+    {:ok, opengl_cpp_internal_state_4k} = Membrane.VideoCompositor.Implementation.OpenGL.Cpp.init(caps_4k)
 
-    {:ok, opengl_rust_internal_state_720p} = Membrane.VideoCompositor.OpenGL.Rust.init(caps_720p)
-    {:ok, opengl_rust_internal_state_1080p} = Membrane.VideoCompositor.OpenGL.Rust.init(caps_1080p)
-    {:ok, opengl_rust_internal_state_4k} = Membrane.VideoCompositor.OpenGL.Rust.init(caps_4k)
+    {:ok, opengl_rust_internal_state_720p} = Membrane.VideoCompositor.Implementation.OpenGL.Rust.init(caps_720p)
+    {:ok, opengl_rust_internal_state_1080p} = Membrane.VideoCompositor.Implementation.OpenGL.Rust.init(caps_1080p)
+    {:ok, opengl_rust_internal_state_4k} = Membrane.VideoCompositor.Implementation.OpenGL.Rust.init(caps_4k)
 
     {:ok, nx_internal_state_720p} = Membrane.VideoCompositor.Nx.init(caps_720p)
     {:ok, nx_internal_state_1080p} = Membrane.VideoCompositor.Nx.init(caps_1080p)
@@ -99,11 +99,11 @@ defmodule Membrane.VideoCompositor.Benchmark.MergeFrames do
     Benchee.run(
       %{
         "FFmpeg - Merge two frames to one" =>
-          fn {frames, internal_states} -> Membrane.VideoCompositor.FFmpeg.merge_frames(frames, internal_states.ffmpeg) end,
+          fn {frames, internal_states} -> Membrane.VideoCompositor.Implementation.FFmpeg.merge_frames(frames, internal_states.ffmpeg) end,
         "OpenGL C++ - Merge two frames to one" =>
-          fn {frames, internal_states} -> Membrane.VideoCompositor.OpenGL.Cpp.merge_frames(frames, internal_states.opengl_cpp) end,
+          fn {frames, internal_states} -> Membrane.VideoCompositor.Implementation.OpenGL.Cpp.merge_frames(frames, internal_states.opengl_cpp) end,
         "OpenGL Rust - Merge two frames to one" =>
-          fn {frames, internal_states} -> Membrane.VideoCompositor.OpenGL.Rust.merge_frames(frames, internal_states.opengl_rust) end,
+          fn {frames, internal_states} -> Membrane.VideoCompositor.Implementation.OpenGL.Rust.merge_frames(frames, internal_states.opengl_rust) end,
         "Nx - Merge two frames to one" =>
           fn {frames, internal_states} -> Membrane.VideoCompositor.Nx.merge_frames(frames, internal_states.nx) end
       },
