+++ c_src/ffmpeg_video_compositor/video_compositor.spec.exs
@@ -1,9 +1,9 @@
-module Membrane.VideoCompositor.FFmpeg.Native
+module Membrane.VideoCompositor.Implementation.FFmpeg.Native
 
 state_type "State"
 
 type(
-  raw_video :: %Membrane.VideoCompositor.FFmpeg.Native.RawVideo{
+  raw_video :: %Membrane.VideoCompositor.Implementation.FFmpeg.Native.RawVideo{
     width: int,
     height: int,
     pixel_format: atom,
