+++ lib/membrane/video_compositor/frame_compositor.ex
