+++ lib/membrane/video_compositor/implementation/ffmpeg/native.ex
@@ -1,9 +1,9 @@
-defmodule Membrane.VideoCompositor.FFmpeg.Native do
+defmodule Membrane.VideoCompositor.Implementation.FFmpeg.Native do
   @moduledoc false
   use Unifex.Loader
 end
 
-defmodule Membrane.VideoCompositor.FFmpeg.Native.RawVideo do
+defmodule Membrane.VideoCompositor.Implementation.FFmpeg.Native.RawVideo do
   @moduledoc """
   Unifex compatible implementation of Membrane.RawVideo struct.
   It should be used only with the native FFmpeg module functions.
