+++ lib/membrane/video_compositor/implementation/opengl/native/cpp.ex
@@ -1,0 +1,4 @@
+defmodule Membrane.VideoCompositor.Implementation.OpenGL.Native.Cpp do
+  @moduledoc false
+  use Unifex.Loader
+end
