+++ lib/membrane/video_compositor/multiple_inputs/video_compositor/multiple_inputs_frame_compositor.ex
@@ -1,4 +1,4 @@
-defmodule Membrane.VideoCompositor.FrameCompositor.MultipleInputs do
+defmodule Membrane.VideoCompositor.MultipleInputs.VideoCompositor.FrameCompositor do
   @moduledoc """
   This module defines behaviour for different frame composition implementations supporting multiple input frames.
   """
