+++ test/opengl/opengl_cpp_native_test.exs
@@ -2,7 +2,7 @@ defmodule VideoCompositor.OpenGL.Cpp.Native.Test do
   use ExUnit.Case
 
   alias Membrane.RawVideo
-  alias Membrane.VideoCompositor.OpenGL.Native.Cpp
+  alias Membrane.VideoCompositor.Implementation.OpenGL.Native.Cpp
   alias Membrane.VideoCompositor.Utility
 
   describe "OpenGL cpp native test on " do
@@ -21,7 +21,9 @@ defmodule VideoCompositor.OpenGL.Cpp.Native.Test do
       }
 
       {:ok, video} =
-        Membrane.VideoCompositor.FFmpeg.Native.RawVideo.from_membrane_raw_video(video)
+        Membrane.VideoCompositor.Implementation.FFmpeg.Native.RawVideo.from_membrane_raw_video(
+          video
+        )
 
       assert in_frame = File.read!(in_path)
