+++ test/opengl/opengl_rust_native_test.exs
@@ -1,7 +1,7 @@
 defmodule VideoCompositor.OpenGL.Rust.Test do
   use ExUnit.Case, async: false
 
-  alias Membrane.VideoCompositor.OpenGL.Native.Rust
+  alias Membrane.VideoCompositor.Implementation.OpenGL.Native.Rust
   alias Membrane.VideoCompositor.Utility
 
   describe "OpenGL rust native test on " do
